refactor: combine FreeTextRequestItemRenderer and DecidableFreeTextRequestItemRenderer
