--- packages/renderers/lib/src/request/request_item_renderer/decidable/decidable_free_text_request_item.dart
@@ -1,56 +1,0 @@
-import 'package:enmeshed_types/enmeshed_types.dart';
-import 'package:enmeshed_ui_kit/enmeshed_ui_kit.dart';
-import 'package:flutter/material.dart';
-
-import '../../../custom_list_tile.dart';
-import '../../request_item_index.dart';
-import '../../request_renderer_controller.dart';
-import 'checkbox_enabled_extension.dart';
-import 'widgets/handle_checkbox_change.dart';
-
-class DecidableFreeTextRequestItemRenderer extends StatefulWidget {
-  final FreeTextRequestItemDVO item;
-  final RequestRendererController? controller;
-  final RequestItemIndex itemIndex;
-  final LocalRequestStatus? requestStatus;
-
-  const DecidableFreeTextRequestItemRenderer({super.key, required this.item, this.controller, required this.itemIndex, this.requestStatus});
-
-  @override
-  State<DecidableFreeTextRequestItemRenderer> createState() => _DecidableFreeTextRequestItemRendererState();
-}
-
-class _DecidableFreeTextRequestItemRendererState extends State<DecidableFreeTextRequestItemRenderer> {
-  late bool isChecked;
-
-  @override
-  void initState() {
-    super.initState();
-
-    isChecked = widget.item.initiallyChecked;
-
-    if (isChecked) {
-      widget.controller?.writeAtIndex(index: widget.itemIndex, value: const AcceptFreeTextRequestItemParameters(freeText: ''));
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Checkbox(value: isChecked, onChanged: widget.item.checkboxEnabled ? onUpdateCheckbox : null),
-        Expanded(child: CustomListTile(title: widget.item.name, description: widget.item.description, thirdLine: widget.item.freeText)),
-      ],
-    );
-  }
-
-  void onUpdateCheckbox(bool? value) {
-    if (value == null) return;
-
-    setState(() {
-      isChecked = value;
-    });
-
-    handleCheckboxChange(isChecked: isChecked, controller: widget.controller, itemIndex: widget.itemIndex);
-  }
-}
