+++ packages/renderers/lib/src/request/request_item_renderer/request_item_renderer.dart
@@ -116,10 +116,12 @@ class RequestItemRenderer extends StatelessWidget {
           dvo.isDecidable
               ? DecidableRegisterAttributeListenerRequestItemRenderer(controller: controller, item: dvo, itemIndex: itemIndex)
               : RegisterAttributeListenerRequestItemRenderer(item: dvo),
-        final FreeTextRequestItemDVO dvo =>
-          dvo.isDecidable
-              ? DecidableFreeTextRequestItemRenderer(controller: controller, item: dvo, itemIndex: itemIndex, requestStatus: requestStatus)
-              : FreeTextRequestItemRenderer(item: dvo),
+        final FreeTextRequestItemDVO dvo => FreeTextRequestItemRenderer(
+          controller: controller,
+          item: dvo,
+          itemIndex: itemIndex,
+          requestStatus: requestStatus,
+        ),
         final TransferFileOwnershipRequestItemDVO dvo => TransferFileOwnershipRequestItemRenderer(
           controller: controller,
           item: dvo,
